# backend/src/main/resources/api-spec.yaml
openapi: 3.0.3
info:
  title: Borsibaar API
  version: 1.0.0
paths:
...
  /products:
    post:
      tags:
        - product
      summary: Create product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        201:
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
  /products/{id}:
    get:
      tags:
        - product
      summary: Get product by ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Got product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    delete:
      tags:
        - product
      summary: Delete product
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        204:
          description: Product deleted
  /organizations:
    post:
      tags:
        - organization
      summary: Create organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
      responses:
        201:
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
    get:
      tags:
        - organization
      summary: Get all organizations
      operationId: getAllOrganizations
      responses:
        200:
          description: Got organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationResponse'
  /organizations/{id}:
    get:
      tags:
        - organization
      summary: Get organization by ID
      operationId: getOrganizationById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Got organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
    put:
      tags:
        - organization
      summary: Update organization
      operationId: updateOrganization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
      responses:
        200:
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
...
  /inventory/product/{productId}:
    get:
      tags:
        - inventory
      summary: Get product inventory
      operationId: getProductInventory
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Got product inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
...
components:
  schemas:
...
    ProductRequest:
      type: object
      required:
        - name
        - currentPrice
        - minPrice
        - maxPrice
        - categoryId
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        currentPrice:
          type: number
          minimum: 0
          exclusiveMinimum: true
        minPrice:
          type: number
          minimum: 0
          exclusiveMinimum: true
        maxPrice:
          type: number
          minimum: 0
          exclusiveMinimum: true
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
...
    OrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        priceIncreaseStep:
          type: number
          minimum: 0.0
        priceDecreaseStep:
          type: number
          minimum: 0.0
...
    InventoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        organizationId:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        description:
          type: string
        basePrice:
          type: number
        minPrice:
          type: number
        maxPrice:
          type: number
        updatedAt:
          type: string
          format: date-time
...
    OrganizationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        priceIncreaseStep:
          type: number
        priceDecreaseStep:
          type: number
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        currentPrice:
          type: number
        minPrice:
          type: number
        maxPrice:
          type: number
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
...